llm-code-deployer/backend/requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
httpx==0.24.1
python-dotenv==1.0.0
PyGithub==2.1.1
openai==1.0.0
requests>=2.31.0
certifi>=2024.7.4
